refactor(megatron): rehome still-used trainer utilities into the backend
Split the legacy megatron trainer/utils into their natural homes in the
megatron backend: PP-visualization wrappers -> core/pipeline_parallel/
pp_visualizer.py; ROCm arg validation -> patches/args/rocm_arg_validation.py;
batch loader (DataLoaderStore/get_batch_func) -> data_loader_store.py;
is_v_schedule_enabled -> training/utils.py; and the manual-loop pretrain
trainer (used by performance projection's layer benchmark) -> the
primus.backends.megatron.pretrainer package. Update all importers. Behavior
preserved.
