+++ primus/backends/megatron/core/pipeline_parallel/primuspipe/handlers/fwd_handler.py
@@ -11,13 +11,15 @@ from megatron.core.pipeline_parallel.schedules import (
 )
 from megatron.training.global_vars import get_args
 
+from primus.backends.megatron.core.pipeline_parallel.pp_visualizer import (
+    fwd_bwd_wrapper,
+)
 from primus.core.pipeline_parallel.handler.offload_handler import OFFLOAD_BUFFER
 from primus.core.pipeline_parallel.scheduler.scheduler_node import (
     FuncType,
     SchedulerNode,
 )
 from primus.core.pipeline_parallel.utils import find_prev_node_with_type
-from primus.modules.trainer.megatron.utils import fwd_bwd_wrapper
 
 
 def megatron_check_fwd_node_valid(node: SchedulerNode):
