+++ primus/backends/megatron/pretrainer/__init__.py
@@ -1,0 +1,5 @@
+###############################################################################
+# Copyright (c) 2025, Advanced Micro Devices, Inc. All rights reserved.
+#
+# See LICENSE for license information.
+###############################################################################
