--- primus/modules/trainer/megatron/utils.py
@@ -1,584 +1,0 @@
-###############################################################################
-# Copyright (c) 2025, Advanced Micro Devices, Inc. All rights reserved.
-#
-# See LICENSE for license information.
-###############################################################################
-
-"""megatron utils"""
-
-import inspect
-import json
-import os
-
-import megatron
-import torch
-from megatron.core import parallel_state
-from megatron.training.global_vars import get_args
-
-from primus.core.utils import logger
-
-_GLOBAL_PP_VIS_EVENTS = []
-_GLOBAL_PP_VIS_EVENTS_PER_ITER = None
-
-
-######################################################log after torch distributed initialized
-
-
-def is_v_schedule_enabled(args=None):
-    if args is None:
-        args = get_args()
-    return (
-        args.patch_zero_bubble
-        and args.enable_zero_bubble
-        and (args.zero_bubble_v_schedule or args.enable_1f1b_v)
-    ) or (args.pp_algorithm in ("zbv-formatted", "v-half", "v-min") and args.patch_primus_pipeline)
-
-
-def is_last_rank():
-    return torch.distributed.get_rank() == (torch.distributed.get_world_size() - 1)
-
-
-def print_rank_last(msg):
-    """If distributed is initialized, print only on last rank."""
-    log_func = logger.info_with_caller
-
-    caller = inspect.stack()[1]
-    caller_frame = caller.frame
-    function_name = caller_frame.f_code.co_name
-    module_name = caller_frame.f_globals["__name__"].split(".")[-1]
-    line = caller.lineno
-
-    if torch.distributed.is_initialized():
-        if is_last_rank():
-            log_func(msg, module_name, function_name, line)
-    else:
-        log_func(msg, module_name, function_name, line)
-
-
-def set_wandb_writer_patch(args):  # monkey patch
-    """
-    This function is adapted from the original Megatron implementation, with an additional
-    wandb argument `entity` be added.
-    Monkey-patch note:
-    - The original function will be replaced at runtime by this implementation.
-
-    """
-
-    megatron.training.global_vars._ensure_var_is_not_initialized(
-        megatron.training.global_vars._GLOBAL_WANDB_WRITER, "wandb writer"
-    )
-
-    if getattr(args, "wandb_project", "") and args.rank == (args.world_size - 1):
-        if args.wandb_exp_name == "":
-            raise ValueError("Please specify the wandb experiment name!")
-
-        import wandb
-
-        if args.wandb_save_dir:
-            save_dir = args.wandb_save_dir
-        else:
-            # Defaults to the save dir.
-            save_dir = os.path.join(args.save, "wandb")
-        wandb_kwargs = {
-            "dir": save_dir,
-            "name": args.wandb_exp_name,
-            "project": args.wandb_project,
-            "entity": args.wandb_entity,
-            "config": vars(args),
-        }
-        os.makedirs(wandb_kwargs["dir"], exist_ok=True)
-        wandb.init(**wandb_kwargs)
-        megatron.training.global_vars._GLOBAL_WANDB_WRITER = wandb
-
-
-def validate_specified_recompute_layers(args):
-    if args.recompute_layer_ids is None:
-        return
-
-    assert isinstance(
-        args.recompute_layer_ids, list
-    ), f"recompute_layer_ids={args.recompute_layer_ids} should be a list"
-    recompute_layer_ids = list(set(args.recompute_layer_ids))
-    assert len(recompute_layer_ids) > 0, "recompute layer ids is null"
-    for layer_id in recompute_layer_ids:
-        assert (
-            layer_id >= 0 and layer_id < args.num_layers
-        ), f"recompute layer id must be between 0 and {args.num_layers - 1}"
-
-    if args.recompute_granularity != "full":
-        raise ValueError(
-            f'When using recompute_layer_ids, recompute_granuarlity: {args.recompute_granularity} must be "full"'
-        )
-
-    if args.recompute_method is not None:
-        raise ValueError(
-            f"When using recompute_layer_ids, recompute_method: {args.recompute_method} must be None."
-        )
-
-    if args.distribute_saved_activations and args.sequence_parallel:
-        raise ValueError(
-            f"distribute_saved_activations: {args.distribute_saved_activations} must be "
-            f"false when sequence parallel is enabled: {args.sequence_parallel}"
-        )
-
-
-def validate_manual_split(args):
-    """
-    The use of decoder_pipeline_manual_split_list is to relax the divisibility
-    restriction of the current (interleaved) 1f1b pipeline schedule. The layer
-    split or number of each pp rank is
-    decoder_pipeline_manual_split_list[pp_rank*vp_size:(pp_rank+1)*vp_size] or
-    decoder_pipeline_manual_split_list[pp_rank] when interleaved pipeline is
-    used or not. For example, the split list could be "[2,3,2,2,2,2,2,1]"
-    in layer16-pp4-vpp2 config, where the vpp split of
-    pp_rank0/pp_rank1/pp_rank2/pp_rank3 is [2,3]/[2,2]/[2,2]/[2,1].
-
-    if chosen pipeline is v_schedule like zbv/v-half,
-    the split list will be the actual layer sequence.
-    For example, layer16-pp4-vpp2 config, the vpp split of
-    pp_rank0/pp_rank1/pp_rank2/pp_rank3 is [3,2,2,2,2,2,2,1]
-    indicate the pipeline as follows:
-    pp_rank0: 3       1
-    pp_rank1:  2     2
-    pp_rank2:   2   2
-    pp_rank3:    2 2
-
-    """
-
-    if (
-        args.num_layers_per_virtual_pipeline_stage is not None
-        or args.decoder_first_pipeline_num_layers is not None
-        or args.decoder_last_pipeline_num_layers is not None
-        or args.account_for_embedding_in_pipeline_split
-        or args.account_for_loss_in_pipeline_split
-    ):
-        raise ValueError(
-            "decoder_pipeline_manual_split_list is not compatible "
-            "with num_layers_per_virtual_pipeline_stage/"
-            "decoder_first_pipeline_num_layers/"
-            "decoder_last_pipeline_num_layers/"
-            "account_for_embedding_in_pipeline_split/"
-            "account_for_loss_in_pipeline_split yet"
-        )
-
-    num_layers = args.num_layers
-    pp_size = args.pipeline_model_parallel_size
-    vp_size = args.virtual_pipeline_model_parallel_size
-    pp_split = args.decoder_pipeline_manual_split_list
-
-    if pp_size <= 1:
-        raise ValueError(
-            f"pipeline_model_parallel_size={pp_size} should be larger "
-            f"than 1 when decoder_pipeline_manual_split_list is used"
-        )
-
-    if not isinstance(pp_split, list):
-        raise ValueError(f"decoder_pipeline_manual_split_list={pp_split} should be a list")
-
-    split_size = pp_size if vp_size is None else pp_size * vp_size
-    if len(pp_split) != split_size:
-        raise ValueError(
-            f"the size of decoder_pipeline_manual_split_list="
-            f"{pp_split} should be {split_size} "
-            f"given pipeline_model_parallel_size={pp_size} and "
-            f"virtual_pipeline_model_parallel_size={vp_size}"
-        )
-
-    if not all(x > 0 for x in pp_split):
-        raise ValueError(
-            f"layer numbers in decoder_pipeline_manual_split_list={pp_split} should all be larger than 0"
-        )
-
-    if sum(pp_split) != num_layers:
-        raise ValueError(
-            f"the sum of decoder_pipeline_manual_split_list="
-            f"{pp_split} is {sum(pp_split)} and "
-            f"should be equal to num_layers={num_layers}"
-        )
-
-    return True
-
-
-def validate_args_modified(*args, **kwargs):
-    def validate_args_modifier(func, modification):
-        import inspect
-
-        source = inspect.getsource(func)
-        modified_source = modification(source)
-        namespace = {}
-        exec(modified_source, func.__globals__, namespace)
-        return namespace[func.__name__]
-
-    ori_code = kwargs.pop("ori_code", None)
-    new_code = kwargs.pop("new_code", None)
-
-    assert ori_code is not None and new_code is not None, "ori_code and new_code must be provided."
-
-    megatron.training.arguments.validate_args = validate_args_modifier(
-        megatron.training.arguments.validate_args, lambda s: s.replace(ori_code, new_code)
-    )
-    megatron.training.arguments.validate_args(*args, **kwargs)
-
-
-def set_manual_pipeline_split_patch(args):
-    """
-    Monkey-patch note:
-    - The original function will be replaced at runtime by this implementation.
-
-    """
-
-    megatron.core.transformer.TransformerConfig.decoder_pipeline_manual_split_list = (
-        args.decoder_pipeline_manual_split_list
-    )
-
-    # patch get_num_layers_to_build
-    def get_num_layers_to_build_patch(config, vp_stage, pp_rank=None):
-        if pp_rank is None:
-            pp_rank = parallel_state.get_pipeline_model_parallel_rank()
-        vp_size = config.virtual_pipeline_model_parallel_size
-
-        if not is_v_schedule_enabled():
-            pp_idx = pp_rank if vp_size is None else pp_rank * vp_size + vp_stage
-            num_layers_to_build = config.decoder_pipeline_manual_split_list[pp_idx]
-            return num_layers_to_build
-        else:
-            assert vp_stage is not None and vp_stage in (0, 1)
-            pp_size = config.pipeline_model_parallel_size
-            chunk_id = pp_rank if vp_stage == 0 else 2 * pp_size - pp_rank - 1
-            num_layers_to_build = config.decoder_pipeline_manual_split_list[chunk_id]
-            return num_layers_to_build
-
-    megatron.core.transformer.transformer_block.get_num_layers_to_build = get_num_layers_to_build_patch
-    megatron.core.models.gpt.gpt_layer_specs.get_num_layers_to_build = get_num_layers_to_build_patch
-
-    # patch get_transformer_layer_offset
-    def get_transformer_layer_offset_patch(config, vp_stage, pp_rank=None):
-        if pp_rank is None:
-            pp_rank = parallel_state.get_pipeline_model_parallel_rank()
-        pp_size = config.pipeline_model_parallel_size
-        vp_size = config.virtual_pipeline_model_parallel_size
-
-        offset = 0
-
-        if not is_v_schedule_enabled():
-            if vp_stage is not None:
-                for vp_idx in range(vp_stage):
-                    for pp_idx in range(pp_size):
-                        offset += config.decoder_pipeline_manual_split_list[pp_idx * vp_size + vp_idx]
-                for pp_idx in range(pp_rank):
-                    offset += config.decoder_pipeline_manual_split_list[pp_idx * vp_size + vp_stage]
-            else:
-                offset = sum(config.decoder_pipeline_manual_split_list[:pp_rank])
-        else:
-            assert vp_stage is not None and vp_stage in (0, 1)
-            chunk_id = pp_rank if vp_stage == 0 else 2 * pp_size - pp_rank - 1
-            offset = sum(config.decoder_pipeline_manual_split_list[:chunk_id])
-        return offset
-
-    megatron.core.transformer.transformer_layer.get_transformer_layer_offset = (
-        get_transformer_layer_offset_patch
-    )
-    megatron.core.transformer.transformer_block.get_transformer_layer_offset = (
-        get_transformer_layer_offset_patch
-    )
-    megatron.core.models.gpt.gpt_layer_specs.get_transformer_layer_offset = get_transformer_layer_offset_patch
-
-
-def schedule_wrapper(func):
-    def wrapper(*args, **kwargs):
-        global _GLOBAL_PP_VIS_EVENTS_PER_ITER
-        _GLOBAL_PP_VIS_EVENTS_PER_ITER = {
-            "start": None,
-            "end": None,
-            "memory": None,
-            "fwd_start": [],
-            "fwd_end": [],
-            "fwd_minibatch": [],
-            "fwd_chunk": [],
-            "bwd_start": [],
-            "bwd_end": [],
-            "bwd_minibatch": [],
-            "bwd_chunk": [],
-            "wgrad_start": [],
-            "wgrad_end": [],
-            "wgrad_minibatch": [],
-            "wgrad_chunk": [],
-        }
-
-        _GLOBAL_PP_VIS_EVENTS_PER_ITER["start"] = torch.cuda.Event(enable_timing=True)
-        _GLOBAL_PP_VIS_EVENTS_PER_ITER["start"].record()
-        res = func(*args, **kwargs)
-        _GLOBAL_PP_VIS_EVENTS_PER_ITER["end"] = torch.cuda.Event(enable_timing=True)
-        _GLOBAL_PP_VIS_EVENTS_PER_ITER["end"].record()
-
-        _GLOBAL_PP_VIS_EVENTS_PER_ITER["memory"] = torch.cuda.max_memory_reserved() / 1024**3
-
-        global _GLOBAL_PP_VIS_EVENTS
-        _GLOBAL_PP_VIS_EVENTS.append(_GLOBAL_PP_VIS_EVENTS_PER_ITER)
-
-        return res
-
-    return wrapper
-
-
-def fwd_bwd_wrapper(func, mode, minibatch=None, chunk=None):
-    def wrapper(*args, **kwargs):
-        start = torch.cuda.Event(enable_timing=True)
-        end = torch.cuda.Event(enable_timing=True)
-
-        start.record()
-        res = func(*args, **kwargs)
-        end.record()
-
-        global _GLOBAL_PP_VIS_EVENTS_PER_ITER
-        _GLOBAL_PP_VIS_EVENTS_PER_ITER[mode + "_start"].append(start)
-        _GLOBAL_PP_VIS_EVENTS_PER_ITER[mode + "_end"].append(end)
-
-        if minibatch is not None:
-            _GLOBAL_PP_VIS_EVENTS_PER_ITER[mode + "_minibatch"].append(minibatch)
-        if chunk is not None:
-            _GLOBAL_PP_VIS_EVENTS_PER_ITER[mode + "_chunk"].append(chunk)
-        return res
-
-    return wrapper
-
-
-def combined_fwd_bwd_wrapper(func, fwd_minibatch, fwd_chunk, bwd_minibatch, bwd_chunk):
-    """Record a single combined forward+backward call as both an ``fwd`` event
-    and a ``bwd`` event sharing the same ``[start, end]`` interval.
-
-    Used by ``megatron_combined_fwd_bkwd_handler`` so that nodes collapsed into
-    a combined FB group still appear in the dump_pp_data output. Without this
-    the visualizer's per-rank F/B/W totals are heavily under-counted on ranks
-    that hit the steady state (the F and B halves are interleaved inside
-    ``combined_forward_backward_step`` and cannot be timed separately).
-    """
-
-    def wrapper(*args, **kwargs):
-        start = torch.cuda.Event(enable_timing=True)
-        end = torch.cuda.Event(enable_timing=True)
-
-        start.record()
-        res = func(*args, **kwargs)
-        end.record()
-
-        global _GLOBAL_PP_VIS_EVENTS_PER_ITER
-        _GLOBAL_PP_VIS_EVENTS_PER_ITER["fwd_start"].append(start)
-        _GLOBAL_PP_VIS_EVENTS_PER_ITER["fwd_end"].append(end)
-        _GLOBAL_PP_VIS_EVENTS_PER_ITER["fwd_minibatch"].append(fwd_minibatch)
-        _GLOBAL_PP_VIS_EVENTS_PER_ITER["fwd_chunk"].append(fwd_chunk)
-        _GLOBAL_PP_VIS_EVENTS_PER_ITER["bwd_start"].append(start)
-        _GLOBAL_PP_VIS_EVENTS_PER_ITER["bwd_end"].append(end)
-        _GLOBAL_PP_VIS_EVENTS_PER_ITER["bwd_minibatch"].append(bwd_minibatch)
-        _GLOBAL_PP_VIS_EVENTS_PER_ITER["bwd_chunk"].append(bwd_chunk)
-        return res
-
-    return wrapper
-
-
-def set_dump_pp_data_patch():
-    from megatron.core.pipeline_parallel import schedules
-
-    schedules.forward_step = fwd_bwd_wrapper(schedules.forward_step, "fwd")
-    schedules.backward_step = fwd_bwd_wrapper(schedules.backward_step, "bwd")
-
-
-def dump_pp_data(args, num_mbs, pp_data_dir):
-    torch.cuda.synchronize()
-
-    global _GLOBAL_PP_VIS_EVENTS
-    all_iter_data = {}
-    for iter_idx, iter_events in enumerate(_GLOBAL_PP_VIS_EVENTS):
-        iter_data = {
-            "total": None,
-            "memory": None,
-            "fwd_start": [],
-            "fwd_end": [],
-            "fwd_minibatch": [],
-            "fwd_chunk": [],
-            "bwd_start": [],
-            "bwd_end": [],
-            "bwd_minibatch": [],
-            "bwd_chunk": [],
-            "wgrad_start": [],
-            "wgrad_end": [],
-            "wgrad_minibatch": [],
-            "wgrad_chunk": [],
-        }
-        iter_data["total"] = iter_events["start"].elapsed_time(iter_events["end"])
-        iter_data["memory"] = iter_events["memory"]
-
-        for i in range(len(iter_events["fwd_start"])):
-            for key in ["fwd_start", "fwd_end", "bwd_start", "bwd_end", "wgrad_start", "wgrad_end"]:
-                if i >= len(iter_events[key]):
-                    continue
-                event_time = iter_events["start"].elapsed_time(iter_events[key][i])
-                iter_data[key].append(event_time)
-            for key in [
-                "fwd_minibatch",
-                "fwd_chunk",
-                "bwd_minibatch",
-                "bwd_chunk",
-                "wgrad_minibatch",
-                "wgrad_chunk",
-            ]:
-                if i >= len(iter_events[key]):
-                    continue
-                iter_data[key].append(iter_events[key][i])
-
-        all_iter_data[iter_idx + 1] = iter_data
-
-    rank = torch.distributed.get_rank()
-    dp_rank = parallel_state.get_data_parallel_rank()
-    pp_rank = parallel_state.get_pipeline_model_parallel_rank()
-    os.makedirs(pp_data_dir, exist_ok=True)
-    if dp_rank == 0:
-        log_path = os.path.join(pp_data_dir, f"pp_rank_{pp_rank}.json")
-        with open(log_path, "w") as f:
-            json.dump(all_iter_data, f, indent=2)
-
-    if rank == 0:
-        vp_size = args.virtual_pipeline_model_parallel_size
-        vp_size = 1 if vp_size is None else vp_size
-        config_dict = {
-            "world_size": args.world_size,
-            "dp_size": args.data_parallel_size,
-            "tp_size": args.tensor_model_parallel_size,
-            "ep_size": args.expert_model_parallel_size,
-            "pp_size": args.pipeline_model_parallel_size,
-            "vp_size": vp_size,
-            "num_mbs": num_mbs,
-            "train_iters": args.train_iters,
-        }
-        log_path = os.path.join(pp_data_dir, f"config.json")
-        with open(log_path, "w") as f:
-            json.dump(config_dict, f, indent=2)
-
-
-def _get_sync_free_moe_options(args) -> dict:
-    stage = args.turbo_sync_free_moe_stage
-
-    if stage > 3 or stage < 0:
-        raise ValueError("turbo_sync_free_moe_stage only support [0-3]")
-
-    sync_free_moe = {
-        1: {
-            "moe_use_fused_router_with_aux_score": True,
-            "moe_permute_fusion": True,
-            "moe_router_padding_for_quantization": True if args.fp8 or args.fp4 else False,
-        },
-        2: {
-            "moe_use_fused_router_with_aux_score": True,
-            "use_turbo_deepep": True,
-            "moe_permute_fusion": True,
-            "use_turbo_grouped_gemm": True,
-            "moe_router_padding_for_quantization": True if args.fp8 or args.fp4 else False,
-        },
-        3: {
-            "moe_use_fused_router_with_aux_score": True,
-            "use_turbo_deepep": True,
-            "moe_permute_fusion": True,
-            "use_turbo_grouped_gemm": True,
-            "moe_router_padding_for_quantization": True if args.fp8 or args.fp4 else False,
-            "use_turbo_fused_act_with_probs": True,
-        },
-    }
-
-    return sync_free_moe[stage]
-
-
-def validate_args_on_rocm(args):
-    # Deterministic mode
-    if args.deterministic_mode:
-        # NOTE: Some environment variables affect deterministic mode on ROCm. Need to do extra check.
-        NON_DETERMINISTIC_ENVS = {
-            "TORCH_COMPILE_DISABLE": "1",
-            "ROCBLAS_DEFAULT_ATOMICS_MODE": "0",
-            "PRIMUS_TURBO_AUTO_TUNE": "0",
-            "PRIMUS_DETERMINISTIC": "1",
-        }
-        # NOTE: Some version triton compile exist potential racing condition issue.
-        for env, value in NON_DETERMINISTIC_ENVS.items():
-            assert (
-                os.environ.get(env, None) == value
-            ), f"{env} must be set to {value} in deterministic mode but got {os.environ.get(env, None)} instead."
-
-        # Set fill_uninitialized_memory to False to avoid calling extra fill kernel in deterministic mode.
-        torch.utils.deterministic.fill_uninitialized_memory = False
-
-    assert not getattr(
-        args, "use_turbo_parallel_linear", False
-    ), "use_turbo_parallel_linear has been removed; please use use_turbo_gemm instead."
-    use_turbo_gemm = getattr(args, "use_turbo_gemm", False)
-    # Turbo FP8 linear check
-    if args.fp8 and use_turbo_gemm:
-        support_fp8_recipe = ["tensorwise", "blockwise", "mxfp8"]
-        assert (
-            args.fp8_recipe in support_fp8_recipe
-        ), f"{args.fp8_recipe} recipe is not support when enable `use_turbo_gemm`."
-
-    # Turbo FP4 linear check
-    if args.fp4 and use_turbo_gemm:
-        support_fp4_recipe = ["mxfp4"]
-        assert (
-            args.fp4_recipe in support_fp4_recipe
-        ), f"{args.fp4_recipe} recipe is not support when enable `use_turbo_gemm`."
-
-    # NOTE: mxfp8 environment variable must be set to 1 to enable mxfp8 recipe on ROCm.
-    if args.fp8_recipe == "mxfp8":
-        assert (
-            os.getenv("NVTE_ROCM_ENABLE_MXFP8", "0") == "1"
-        ), "Please set `NVTE_ROCM_ENABLE_MXFP8=1` to enable `mxfp8` recipe."
-
-    # dump pp data
-    if args.dump_pp_data and args.pipeline_model_parallel_size == 1:
-        args.dump_pp_data = False
-        print_rank_last(f"Disable args.dump_pp_data since args.pipeline_model_parallel_size=1")
-
-    # PrimusTurboGroupedMLP no longer depends on legacy GroupedMLP; the two
-    # flags are mutually exclusive when turbo is enabled.
-    assert not getattr(
-        args, "use_turbo_grouped_mlp", False
-    ), "use_turbo_grouped_mlp has been removed; please use use_turbo_grouped_gemm instead."
-    use_turbo_grouped_gemm = getattr(args, "use_turbo_grouped_gemm", False)
-    if use_turbo_grouped_gemm:
-        if getattr(args, "moe_use_legacy_grouped_gemm", False):
-            raise ValueError(
-                "use_turbo_grouped_gemm=True is incompatible with moe_use_legacy_grouped_gemm=True. "
-                "please set moe_use_legacy_grouped_gemm=False."
-            )
-
-    # sync-free MoE
-    if args.turbo_sync_free_moe_stage > 0:
-        assert args.enable_primus_turbo, "Please set `enable_primus_turbo=True` to enable sync-free MoE."
-
-        if args.turbo_sync_free_moe_stage > 1 and not use_turbo_grouped_gemm:
-            raise ValueError(
-                "Sync-Free MoE stage 2 or 3 require PrimusTurboGroupedLinear, please set `use_turbo_grouped_gemm=True`"
-            )
-        options = _get_sync_free_moe_options(args)
-        print_rank_last(
-            f"========== Enable Sync-Free MoE Stage {args.turbo_sync_free_moe_stage} (Auto-Enabled Options) =========="
-        )
-        for flag, value in options.items():
-            dots = "." * (73 - len(flag) - len(str(value)))
-            print_rank_last(f"{flag}{dots}{value}")
-            setattr(args, flag, value)
-        print_rank_last(
-            f"========== Enable Sync-Free MoE Stage {args.turbo_sync_free_moe_stage} (Auto-Enabled Options) =========="
-        )
-
-    # turbo deepep
-    if args.use_turbo_deepep:
-        assert (
-            not args.moe_shared_expert_overlap
-        ), "DeepEP not support moe_shared_expert_overlap, please set `moe_shared_expert_overlap=False`."
-        assert (
-            args.moe_router_dtype == "fp32"
-        ), "DeepEP only supports float32 probs, please set `moe_router_dtype=fp32`"
-        if (
-            args.expert_model_parallel_size >= 16
-            and os.getenv("PRIMUS_TURBO_MOE_DISPATCH_COMBINE_BACKEND", "TURBO") == "TURBO"
-        ):
-            # Turbo DeepEP is not supported for CUs > 32 when using internode dispatch/combine.
-            assert args.turbo_deepep_num_cu <= 32, "Set `turbo_deepep_num_cu<=32` when using ep_size >= 16."
